fix: id

+++ app/models/Message.scala
@@ -45,8 +45,8 @@ class MessageRepository @Inject()(config: Configuration)(implicit ec: ExecutionC
   }
 
   // https://railsguides.jp/active_record_basics.html#crud-%E3%83%87%E3%83%BC%E3%82%BF%E3%81%AE%E8%AA%AD%E3%81%BF%E6%9B%B8%E3%81%8D
-  def update(id: Int, text: String): Future[Int] = {
-    val query = messages.filter(_.id === id.toLong).map(_.text).update(text)
+  def update(id: Long, text: String): Future[Int] = {
+    val query = messages.filter(_.id === id).map(_.text).update(text)
     db.run(query)
   }
 }
